+++ packages/badge/src/index.tsx
@@ -1,10 +1,13 @@
-import * as ReactModule from 'react';
-import { createComponent } from '@lit-labs/react';
+import { createComponent } from '@equinor/fusion-react-utils';
 import HTMLBadgeCustomElement, { tag } from '@equinor/fusion-wc-badge';
 
 export { HTMLBadgeCustomElement };
 
-export const Badge = createComponent(ReactModule, tag, HTMLBadgeCustomElement);
+export type ElementProps = React.PropsWithChildren<
+  Partial<Pick<HTMLBadgeCustomElement, 'size' | 'position' | 'color' | 'value' | 'icon' | 'circular' | 'tooltip'>>
+>;
+
+export const Badge = createComponent<HTMLBadgeCustomElement, ElementProps>(HTMLBadgeCustomElement, tag);
 export type BadgeProps = React.ComponentProps<typeof Badge>;
 
 export default Badge;