+++ packages/date/src/DateTime.tsx
@@ -1,7 +1,15 @@
-import * as React from 'react';
-import { createComponent } from '@lit-labs/react';
+import { createComponent } from '@equinor/fusion-react-utils';
 import { DateTimeElement as HTMLDateTimeCustomElement, dateTimeTag } from '@equinor/fusion-wc-date';
 
-export const DateTime = createComponent(React, dateTimeTag, HTMLDateTimeCustomElement);
 export { HTMLDateTimeCustomElement };
+
+type ElementProps = React.PropsWithChildren<Partial<Pick<HTMLDateTimeCustomElement, 'date' | 'format' | 'locale'>>>;
+
+export const DateTime = createComponent<HTMLDateTimeCustomElement, ElementProps>(
+  HTMLDateTimeCustomElement,
+  dateTimeTag
+);
+
+export type DateTimeProps = React.ComponentProps<typeof DateTime>;
+
 export default DateTime;