+++ storybook/src/stories/switch/Switch.template.tsx
@@ -1,5 +1,0 @@
-import { Switch, SwitchProps } from '@equinor/fusion-react-switch';
-
-export const RadioTemplate = (props: SwitchProps): JSX.Element => <Switch {...props} />;
-
-export default RadioTemplate;