+++ packages/filterpane/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.1.14 (2021-10-29)
+
+**Note:** Version bump only for package @equinor/fusion-react-filter
+
+
+
+
+
+## [1.1.13](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-filter@1.1.12...@equinor/fusion-react-filter@1.1.13) (2021-10-29)
+
+**Note:** Version bump only for package @equinor/fusion-react-filter
+
+
+
+
+
+## 1.1.12 (2021-10-29)
+
+**Note:** Version bump only for package @equinor/fusion-react-filter
+
+
+
+
+
 ## [1.1.11](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-filter@1.1.10...@equinor/fusion-react-filter@1.1.11) (2021-10-18)