+++ packages/skeleton/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.2.1 (2021-10-29)
+
+
+### Bug Fixes
+
+* **skeleton:** add missing prop for icon ([d59e6e0](https://github.com/equinor/fusion-react-components/commit/d59e6e018044cb394245607a05da2d066deb8076))
+
+
+
+
+
+# 0.2.0 (2021-10-29)
+
+
+### Features
+
+* **skeleton:** add new builder ([c900967](https://github.com/equinor/fusion-react-components/commit/c90096718d477eeb3a6948d3fa34422ea0983e1d))
+
+
+
+
+
 ## 0.1.1 (2021-10-27)