+++ packages/switch/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-switch@0.4.0...@equinor/fusion-react-switch@0.4.1) (2021-10-29)
+
+**Note:** Version bump only for package @equinor/fusion-react-switch
+
+
+
+
+
+# 0.4.0 (2021-10-29)
+
+
+### Features
+
+* add new builder ([3a7b239](https://github.com/equinor/fusion-react-components/commit/3a7b239d5e2ecfca05d817f3f99c63ddb4fd1395))
+
+
+
+
+
 ## 0.3.7 (2021-10-13)
 
 **Note:** Version bump only for package @equinor/fusion-react-switch