+++ storybook/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.30.1](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.30.0...@equinor/fusion-react-components-stories@1.30.1) (2021-10-29)
+
+**Note:** Version bump only for package @equinor/fusion-react-components-stories
+
+
+
+
+
+# [1.30.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.29.0...@equinor/fusion-react-components-stories@1.30.0) (2021-10-29)
+
+
+### Features
+
+* add new builder ([3a7b239](https://github.com/equinor/fusion-react-components/commit/3a7b239d5e2ecfca05d817f3f99c63ddb4fd1395))
+
+
+
+
+
+# [1.29.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.28.1...@equinor/fusion-react-components-stories@1.29.0) (2021-10-29)
+
+
+### Features
+
+* **skeleton:** add new builder ([c900967](https://github.com/equinor/fusion-react-components/commit/c90096718d477eeb3a6948d3fa34422ea0983e1d))
+
+
+
+
+
+## [1.28.1](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.28.0...@equinor/fusion-react-components-stories@1.28.1) (2021-10-29)
+
+**Note:** Version bump only for package @equinor/fusion-react-components-stories
+
+
+
+
+
+# [1.28.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.27.0...@equinor/fusion-react-components-stories@1.28.0) (2021-10-28)
+
+
+### Features
+
+* **date:** use internal element builder ([e0a3ac7](https://github.com/equinor/fusion-react-components/commit/e0a3ac790b143c658e697af8e010be7c417308a5))
+
+
+
+
+
+# [1.27.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.26.0...@equinor/fusion-react-components-stories@1.27.0) (2021-10-28)
+
+
+### Features
+
+* **avatar:** use internal element buider ([bd46edf](https://github.com/equinor/fusion-react-components/commit/bd46edfe7ab1116b91ab000c7542c9ae8c5b8f24))
+* **badge:** use internal element buider ([0ae7675](https://github.com/equinor/fusion-react-components/commit/0ae7675846bdb815a235bcf4061ee0de12bb2d94))
+
+
+
+
+
+# [1.26.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.25.2...@equinor/fusion-react-components-stories@1.26.0) (2021-10-28)
+
+
+### Features
+
+* **avatar:** use internal element buider ([bd46edf](https://github.com/equinor/fusion-react-components/commit/bd46edfe7ab1116b91ab000c7542c9ae8c5b8f24))
+
+
+
+
+
 ## [1.25.2](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-components-stories@1.25.1...@equinor/fusion-react-components-stories@1.25.2) (2021-10-27)